pyproject.toml
[project]
name = "tilepygame"
version = "0.1.0"
description = "A Pygame-ce wrapper for top-down 2D tile-based games"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["pygame-ce", "pytmx"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tilepygame"]